+++ _config/task.sass.js
@@ -1,6 +1,24 @@
-/* jshint node: true */
-
 'use strict';
+
+// PostCss plugins
+const autoprefixer = require('autoprefixer');
+const cssNano = require('cssnano');
+const pxToRem = require('postcss-pxtorem');
+const mqPacker = require('css-mqpacker');
+const minifySelectors = require('postcss-minify-selectors');
+
+const postCssConfig = {
+	autoprefixer: {
+		browsers: ['>5%'],
+	},
+	mqpacker: {
+		sort: true,
+	},
+	pxtorem: {
+		replace: false,
+		rootValue: 16,
+	}
+};
 
 function getTaskConfig(projectConfig) {
 
@@ -20,39 +38,47 @@ function getTaskConfig(projectConfig) {
 					projectConfig.paths.src.sass + '/_objects/*.scss',
 					projectConfig.paths.src.sass + '/_components/*.scss',
 					projectConfig.paths.src.components + '**/*.scss',
-					projectConfig.paths.src.sass + '/_trumps/*.scss'
+					projectConfig.paths.src.sass + '/_trumps/*.scss',
 				]
 			}
 		},
+		stylelint: {
+			syntax: 'scss',
+			failAfterError: false,
+			reporters: [
+				{
+					formatter: 'string',
+					save: './.stylelint.log',
+					console: false,
+				}
+			]
+		},
 		watch: [
 			projectConfig.paths.src.sass + '**/*.scss',
-			projectConfig.paths.src.components + '**/*.scss'
+			projectConfig.paths.src.components + '**/*.scss',
 		],
-		defaultConfig: {
-			autoprefixer: {
-				browsers: ['>5%']
-			},
-			mqpacker: {
-				sort: true
-			},
-			pxtorem: {
-				replace:   false,
-				rootValue: 16
-			},
-			stylelint: {
-				syntax: 'scss',
-				failAfterError: false,
-				reporters: [
-					{
-						formatter: 'string',
-						save: './.stylelint.log',
-						console: false
-					}
-				]
+		getPostCssPlugins
+	};
+
+	function getPostCssPlugins() {
+
+		var postCssPlugins = [
+			autoprefixer(postCssConfig.autoprefixer),
+			pxToRem(postCssConfig.pxtorem),
+			mqPacker(postCssConfig.mqpacker),
+			minifySelectors(),
+		];
+
+		if (projectConfig.isProd) {
+			if (postCssConfig.cssNano) {
+				postCssPlugins.push(cssNano(postCssConfig.cssNano));
+			} else {
+				postCssPlugins.push(cssNano());
 			}
 		}
 
-	};
+		return postCssPlugins;
+	}
 
 	return taskConfig;
 }
