+++ package.json
@@ -55,7 +55,6 @@
     "gulp-sass-generate-contents": "~0.3.2",
     "gulp-sourcemaps": "~2.6.0",
     "gulp-stylelint": "^5.0.0",
-    "merge": "^1.2.0",
     "postcss-minify-selectors": "^2.0.5",
     "postcss-pxtorem": "~4.0.0",
     "stylelint-config-standard": "^17.0.0"
